[Add] 完了タスクをTableViewに表示

+++ toDO/Models/Entities/Task.swift
@@ -8,5 +8,5 @@
 
 struct Task {
     var name: String
-    var progress: Int
+    var progress: Float
 }
